MM-42542 Move types/actions back to mattermost-redux
